Revised scheme for marking archived and good data.
Rather than computing everything on the fly, which is too slow with a big
database, we now store the information in the database directly. This means
queries are fast. However, it must be updated after every insertion -- thus,
it is important to call save() and not bulk_create() or update() in the
general case.